+++ scripts/sql_s33_batch.sql
@@ -1,0 +1,88 @@
+BEGIN;
+
+-- Garoumele (id=1012) - 1200-1480 Kanem-Bornu transitional
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(1012, 'Lange, D. (1984). The kingdoms and peoples of Chad. In: D. T. Niane (ed.), General History of Africa, Vol. IV: Africa from the Twelfth to the Sixteenth Century, pp. 238-265. UNESCO/Heinemann. ISBN 978-9231017100.', 'https://unesdoc.unesco.org/ark:/48223/pf0000184282', 'academic'),
+(1012, 'Smith, A. (1971). The early states of the Central Sudan. In: J. F. A. Ajayi & M. Crowder (eds.), History of West Africa, Vol. 1. Longman. ISBN 978-0582646834.', 'https://www.worldcat.org/title/history-of-west-africa/oclc/12046691', 'academic'),
+(1012, 'Cohen, R. (1966). The Dynamics of Feudalism in Bornu. Boston University African Studies Center Papers, 2. Boston University.', 'https://www.bu.edu/africa/publications', 'academic'),
+(1012, 'Hiribarren, V. (2017). A History of Borno: Trans-Saharan African Empire to Failing Nigerian State. Hurst & Co. ISBN 978-1849045643.', 'https://www.hurstpublishers.com/book/a-history-of-borno/', 'academic'),
+(1012, 'Dewière, R. (2017). Du lac Tchad à La Mecque: Le sultanat du Borno et son monde (XVIe-XVIIe siècle). Éditions de la Sorbonne. ISBN 978-2859449698.', 'https://www.editionsdelasorbonne.fr', 'academic');
+
+-- Namayan (id=882) - 1175-1571 Philippine pre-Hispanic
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(882, 'Scott, W. H. (1994). Barangay: Sixteenth-Century Philippine Culture and Society. Ateneo de Manila University Press. ISBN 978-9715501354.', 'https://www.ateneo.edu/aupress', 'academic'),
+(882, 'Huerta, F. de (1865). Estado Geografico, Topografico, Estadistico, Historico-Religioso de la Santa y Apostolica Provincia de San Gregorio Magno. Imprenta de M. Sanchez. (Binondo, Manila)', 'https://www.worldcat.org/title/estado-geografico/oclc/10145687', 'primary'),
+(882, 'Dery, L. C. (2001). A History of the Inarticulate: Local History, Prostitution, and Other Views from the Bottom. New Day Publishers. ISBN 978-9711010690.', 'https://www.worldcat.org/title/history-of-the-inarticulate/oclc/47655097', 'academic'),
+(882, 'Odal-Devora, G. (2000). The River Dwellers. In: R. Vicente Reyes (ed.), Book Pasig: The River of Life, pp. 43-66. Unilever Philippines.', 'https://www.worldcat.org/title/book-pasig-the-river-of-life/oclc/49223100', 'academic'),
+(882, 'Fox, R. B. & Legaspi, A. M. (1977). Excavations at Santa Ana. National Museum of the Philippines.', 'https://www.nationalmuseum.gov.ph', 'academic');
+
+-- Balanguingui (id=905) - 1300-1848 Sulu archipelago piracy
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(905, 'Warren, J. F. (1981). The Sulu Zone, 1768-1898: The Dynamics of External Trade, Slavery, and Ethnicity in the Transformation of a Southeast Asian Maritime State. NUS Press. ISBN 978-9971693862.', 'https://nuspress.nus.edu.sg/products/the-sulu-zone-1768-1898', 'academic'),
+(905, 'Warren, J. F. (2002). Iranun and Balangingi: Globalization, Maritime Raiding and the Birth of Ethnicity. NUS Press. ISBN 978-9971692421.', 'https://nuspress.nus.edu.sg/products/iranun-and-balangingi', 'academic'),
+(905, 'Tarling, N. (1963). Piracy and Politics in the Malay World: A Study of British Imperialism in Nineteenth-Century South-East Asia. F. W. Cheshire / Donald Moore.', 'https://www.worldcat.org/title/piracy-and-politics-in-the-malay-world/oclc/1196961', 'academic'),
+(905, 'Donoso, I. (2023). Bichara: Moro Chanceries and Jawi Legacy in the Philippines. Springer. ISBN 978-9811908200.', 'https://link.springer.com/book/10.1007/978-981-19-0821-7', 'academic'),
+(905, 'Majul, C. A. (1973). Muslims in the Philippines (2nd ed.). University of the Philippines Press. ISBN 978-9711050017.', 'https://up.edu.ph/', 'academic');
+
+-- Etalwa / Lucayan (id=774) - 1000-1550 Bahamas Taíno
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(774, 'Keegan, W. F. (1992). The People Who Discovered Columbus: The Prehistory of the Bahamas. University Press of Florida. ISBN 978-0813011370.', 'https://upf.com/book.asp?id=KEEGA001', 'academic'),
+(774, 'Granberry, J. & Vescelius, G. S. (2004). Languages of the Pre-Columbian Antilles. University of Alabama Press. ISBN 978-0817351236.', 'https://www.uapress.ua.edu/9780817351236/', 'academic'),
+(774, 'Sauer, C. O. (1966). The Early Spanish Main. University of California Press. ISBN 978-0520014152.', 'https://www.ucpress.edu/book/9780520014152', 'academic'),
+(774, 'Keegan, W. F. & Carlson, L. A. (2008). Talking Taíno: Caribbean Natural History from a Native Perspective. Fire Ant Books / University of Alabama Press. ISBN 978-0817355081.', 'https://www.uapress.ua.edu/9780817355081/', 'academic'),
+(774, 'Craton, M. & Saunders, G. (1992). Islanders in the Stream: A History of the Bahamian People, Vol. 1. University of Georgia Press. ISBN 978-0820323527.', 'https://ugapress.org/book/9780820323527/islanders-in-the-stream/', 'academic');
+
+-- Tekrur pre-Almoravid (id=1026) - 500-1040
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(1026, 'Levtzion, N. (1973). Ancient Ghana and Mali. Methuen. ISBN 978-0416758306.', 'https://www.routledge.com/9780416758306', 'academic'),
+(1026, 'Kane, O. (2004). La première hégémonie peule: Le Fuuta Tooro de Koli Teηella à Almaami Abdul. Karthala. ISBN 978-2845863279.', 'https://www.karthala.com', 'academic'),
+(1026, 'Thiaw, I. (2013). From the Senegal River to Siin: Histories of Senegambian Polities and Their Interaction with Outsiders. In: U. Bosma, G. Kessler & L. Lucassen (eds.), Migration and Membership Regimes in Global and Historical Perspective. Brill. ISBN 978-9004241831.', 'https://brill.com/view/title/24211', 'academic'),
+(1026, 'Brooks, G. E. (1985). Western African to c. 1860 A.D.: A Provisional Historical Schema based on Climate Periods. Indiana University African Studies Program.', 'https://africa.indiana.edu', 'academic'),
+(1026, 'Ba, A. (2017). Histoire et politique dans la vallée du fleuve Sénégal. L''Harmattan. ISBN 978-2343121192.', 'https://www.editions-harmattan.fr', 'academic');
+
+-- Niue Patuiki (id=751) - 1700-1900
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(751, 'Smith, S. P. (1903). Niuē-fekai, or Savage Island, and Its People. Whitcombe & Tombs. (Repr. Government of Niue 1983)', 'https://nzetc.victoria.ac.nz/tm/scholarly/tei-SmiNiue.html', 'academic'),
+(751, 'Loeb, E. M. (1926). History and Traditions of Niue. Bernice P. Bishop Museum Bulletin 32. Honolulu.', 'https://www.bishopmuseum.org', 'academic'),
+(751, 'Pollock, N. J. (1979). Work, Food and Drink in Niue. PhD diss., University of California Berkeley.', 'https://escholarship.org', 'academic'),
+(751, 'Chapman, T. M. (1976). The Decolonisation of Niue. New Zealand Institute of International Affairs / Victoria University Press. ISBN 978-0908607013.', 'https://www.vup.victoria.ac.nz', 'academic'),
+(751, 'Talagi, T. (2000). Niue: A History of the Island. Institute of Pacific Studies, University of the South Pacific. ISBN 978-9820201422.', 'https://www.usp.ac.fj', 'academic');
+
+-- Tokelau (id=752) - 1000-1889
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(752, 'Hooper, A. & Huntsman, J. (1996). Tokelau: A Historical Ethnography. University of Hawai''i Press. ISBN 978-0824819125.', 'https://uhpress.hawaii.edu/title/tokelau/', 'academic'),
+(752, 'Maude, H. E. (1981). Slavers in Paradise: The Peruvian Slave Trade in Polynesia, 1862-1864. Australian National University Press. ISBN 978-0708106907.', 'https://press.anu.edu.au', 'academic'),
+(752, 'Angelo, T. & Pasikale, T. (2008). Tokelau: A History of Government. Council for the Ongoing Government of Tokelau / Victoria University Press. ISBN 978-0473131326.', 'https://www.tokelau.org.nz', 'academic'),
+(752, 'Macgregor, G. (1937). Ethnology of Tokelau Islands. Bernice P. Bishop Museum Bulletin 146. Honolulu.', 'https://www.bishopmuseum.org', 'academic'),
+(752, 'Huntsman, J. & Hooper, A. (1976). The Desecration of Tokelau Kinship. Journal of the Polynesian Society, 85(3), 257-273.', 'https://www.thepolynesiansociety.org/jps/volume/85', 'academic');
+
+-- Butaritari (id=769) - 1500-1892 Kiribati
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(769, 'Grimble, A. (1989). Tungaru Traditions: Writings on the Atoll Culture of the Gilbert Islands (ed. H. E. Maude). University of Hawai''i Press / Pacific Islands Monograph Series 7. ISBN 978-0824811785.', 'https://uhpress.hawaii.edu/title/tungaru-traditions/', 'academic'),
+(769, 'Macdonald, B. (2001). Cinderellas of the Empire: Towards a History of Kiribati and Tuvalu. Institute of Pacific Studies, University of the South Pacific. ISBN 978-9820203358.', 'https://www.usp.ac.fj', 'academic'),
+(769, 'Lambert, B. (1971). The Gilbert Islands: Micro-Individualism. In: R. Crocombe (ed.), Land Tenure in the Pacific. Oxford University Press.', 'https://global.oup.com', 'academic'),
+(769, 'Maude, H. E. (1963). The Evolution of the Gilbertese Boti: An Ethnohistorical Interpretation. Polynesian Society Memoir 35.', 'https://www.thepolynesiansociety.org', 'academic'),
+(769, 'Sabatier, E. (1977). Astride the Equator: An Account of the Gilbert Islands (transl. U. Nixon). Oxford University Press. ISBN 978-0195503555.', 'https://global.oup.com', 'academic');
+
+-- Kuhikugu (id=795) - 1000-1600 Upper Xingu Amazonia
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(795, 'Heckenberger, M. J. (2005). The Ecology of Power: Culture, Place, and Personhood in the Southern Amazon, A.D. 1000-2000. Routledge / Psychology Press. ISBN 978-0415945998.', 'https://www.routledge.com/9780415945998', 'academic'),
+(795, 'Heckenberger, M. J., Petersen, J. B. & Neves, E. G. (1999). Village size and permanence in Amazonia: Two archaeological examples from Brazil. Latin American Antiquity, 10(4), 353-376. DOI:10.2307/972028.', 'https://doi.org/10.2307/972028', 'academic'),
+(795, 'Heckenberger, M. J., Kuikuro, A., Kuikuro, U. T., Russell, J. C., Schmidt, M., Fausto, C. & Franchetto, B. (2003). Amazonia 1492: Pristine Forest or Cultural Parkland? Science, 301(5640), 1710-1714. DOI:10.1126/science.1086112.', 'https://doi.org/10.1126/science.1086112', 'academic'),
+(795, 'Heckenberger, M. J., Russell, J. C., Fausto, C., Toney, J. R., Schmidt, M. J., Pereira, E., Franchetto, B. & Kuikuro, A. (2008). Pre-Columbian Urbanism, Anthropogenic Landscapes, and the Future of the Amazon. Science, 321(5893), 1214-1217. DOI:10.1126/science.1159769.', 'https://doi.org/10.1126/science.1159769', 'academic'),
+(795, 'Mann, C. C. (2005). 1491: New Revelations of the Americas Before Columbus. Alfred A. Knopf / Vintage. ISBN 978-1400032051.', 'https://www.penguinrandomhouse.com', 'academic');
+
+-- Amu / Sahrawi (id=826) - 1370-1895
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(826, 'Hodges, T. (1983). Western Sahara: The Roots of a Desert War. Lawrence Hill Books. ISBN 978-0882081526.', 'https://www.worldcat.org/title/western-sahara/oclc/9783604', 'academic'),
+(826, 'Mercer, J. (1976). Spanish Sahara. George Allen & Unwin. ISBN 978-0049660137.', 'https://www.worldcat.org/title/spanish-sahara/oclc/2241540', 'academic'),
+(826, 'Norris, H. T. (1986). The Arab Conquest of the Western Sahara: Studies of the Historical Events, Religious Beliefs and Social Customs Which Made the Remotest Sahara a Part of the Arab World. Longman. ISBN 978-0582756434.', 'https://www.worldcat.org/title/arab-conquest-of-the-western-sahara/oclc/13731113', 'academic'),
+(826, 'Jensen, E. (2005). Western Sahara: Anatomy of a Stalemate. Lynne Rienner. ISBN 978-1588263056.', 'https://www.rienner.com', 'academic'),
+(826, 'Caro Baroja, J. (1955). Estudios Saharianos. Instituto de Estudios Africanos, CSIC. (Reprint Júcar 1990. ISBN 978-8433400956)', 'https://www.csic.es', 'academic');
+
+-- Confidence score upgrades
+UPDATE geo_entities SET confidence_score = 0.85 WHERE id IN (1012, 882, 905, 774, 1026, 751, 752, 769, 795, 826);
+
+COMMIT;
+
+SELECT g.id, g.name_original, g.confidence_score, (SELECT count(*) FROM sources s WHERE s.entity_id=g.id) AS n_src FROM geo_entities g WHERE g.id IN (751, 752, 769, 774, 795, 826, 882, 905, 1012, 1026) ORDER BY g.id;